ml/editing_service/requirements.txt
git+https://github.com/huggingface/diffusers
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.25.0